+++ packages/suite-analytics/CHANGELOG.md
@@ -12,6 +12,7 @@ Renamed:
 
 -   wallet/add-account to accounts/new-account
 -   settings/coin-backend to settings/coins/backend
+-   menu/toggle-onion-links to settings/tor/onion-links
 
 Removed: